refactor(trading-signals): use singular .type.ts suffix and document type conventions
Aligns the pure-type file suffix with .test.ts (singular over plural)
and codifies the conventions in CLAUDE.md: closed type aliases for data
shapes, interface only for contracts that classes implement.
