+++ packages/trading-signals/src/base/BandsResult.type.ts
