+++ packages/trading-signals/src/base/Candle.type.ts
