+++ packages/trading-signals/src/base/index.ts
@@ -1,4 +1,4 @@
-export * from './BandsResult.types.js';
-export * from './Candle.types.js';
+export * from './BandsResult.type.js';
+export * from './Candle.type.js';
 export * from './Indicator.js';
 export * from './Period.js';
