+++ packages/trading-signals/src/momentum/AO/AO.ts
@@ -1,7 +1,7 @@
 import type {MovingAverage} from '../../trend/MA/MovingAverage.js';
 import type {MovingAverageTypes} from '../../trend/MA/MovingAverageTypes.js';
 import {SMA} from '../../trend/SMA/SMA.js';
-import type {HighLow} from '../../base/Candle.types.js';
+import type {HighLow} from '../../base/Candle.type.js';
 import {TrendIndicatorSeries, TradingSignal} from '../../base/Indicator.js';
 
 /**
