+++ packages/trading-signals/src/momentum/WILLR/WilliamsR.ts
@@ -1,5 +1,5 @@
 import {TradingSignal, TrendIndicatorSeries} from '../../base/Indicator.js';
-import type {HighLowClose} from '../../base/Candle.types.js';
+import type {HighLowClose} from '../../base/Candle.type.js';
 import {pushUpdate} from '../../util/pushUpdate.js';
 
 /**
