+++ packages/trading-signals/src/trend/ADX/ADX.ts
@@ -2,7 +2,7 @@ import {DX} from '../DX/DX.js';
 import {IndicatorSeries} from '../../base/Indicator.js';
 import type {MovingAverage} from '../MA/MovingAverage.js';
 import type {MovingAverageTypes} from '../MA/MovingAverageTypes.js';
-import type {HighLowClose} from '../../base/Candle.types.js';
+import type {HighLowClose} from '../../base/Candle.type.js';
 import {WSMA} from '../WSMA/WSMA.js';
 
 /**
