+++ packages/trading-signals/src/trend/BREAKOUT_BAR_LOW/BreakoutBarLow.ts
@@ -1,5 +1,5 @@
 import {IndicatorSeries} from '../../base/Indicator.js';
-import type {HighLow} from '../../base/Candle.types.js';
+import type {HighLow} from '../../base/Candle.type.js';
 import {pushUpdate} from '../../util/pushUpdate.js';
 
 export type BreakoutBarLowConfig = {
