+++ packages/trading-signals/src/volume/AD/AD.ts
@@ -1,4 +1,4 @@
-import type {HighLowCloseVolume} from '../../base/Candle.types.js';
+import type {HighLowCloseVolume} from '../../base/Candle.type.js';
 import {TradingSignal, TrendIndicatorSeries} from '../../base/Indicator.js';
 
 /**
