+++ packages/trading-signals/src/volume/PVT/PVT.ts
@@ -1,4 +1,4 @@
-import type {HighLowCloseVolume} from '../../base/Candle.types.js';
+import type {HighLowCloseVolume} from '../../base/Candle.type.js';
 import {TradingSignal, TrendIndicatorSeries} from '../../base/Indicator.js';
 import {pushUpdate} from '../../util/pushUpdate.js';
